+++ lib/unrouter.dart
@@ -1,1 +1,14 @@
-export 'core.dart';
+export 'src/core/route_data.dart';
+export 'src/core/route_definition.dart';
+export 'src/runtime/unrouter.dart';
+export 'src/runtime/navigation.dart'
+    show
+        UnrouterResolutionState,
+        UnrouterStateSnapshot,
+        UnrouterStateTimelineEntry,
+        UnrouterHistoryStateRequest,
+        UnrouterHistoryStateComposer,
+        UnrouterController,
+        UnrouterControllerLifecycleMethods,
+        UnrouterScope,
+        UnrouterBuildContextExtension;