+++ test/inspector_replay_store_test.dart
@@ -5,6 +5,8 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:unrouter/devtools.dart';
+import 'package:unrouter/machine.dart';
+import 'package:unrouter/unrouter.dart';
 import 'package:unstory/unstory.dart';
 
 void main() {