+++ test/inspector_widget_test.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:unrouter/devtools.dart';
+import 'package:unrouter/unrouter.dart';
 import 'package:unstory/unstory.dart';
 
 void main() {