+++ test/machine_determinism_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:unrouter/machine.dart';
+import 'package:unrouter/unrouter.dart';
 import 'package:unstory/unstory.dart';
 
 void main() {