+++ test/performance_budget_test.dart
@@ -1,5 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:unrouter/devtools.dart';
+import 'package:unrouter/machine.dart';
+import 'package:unrouter/unrouter.dart';
 import 'package:unstory/unstory.dart';
 
 void main() {